requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0.3
django-cleanup==6.0.0
django-cors-headers==3.10.1
django-environ==0.8.1
djangorestframework==3.13.1
Pillow==9.0.1
psycopg2-binary==2.9.3
pyotp==2.6.0
pytz==2022.1
sqlparse==0.4.2
